+++ Tests/TuistKitTests/Commands/FocusCommandTests.swift
@@ -3,7 +3,7 @@ import Foundation
 @testable import TuistCoreTesting
 @testable import TuistKit
 import Utility
-@testable import xcodeproj
+import xcodeproj
 import XCTest
 
 final class FocusCommandTests: XCTestCase {